add router for mapview
